==> Rcode-for-database/00-preamble.R <==
# packages
library(rvest)
library(tidyverse)
library(stringr)
library(readxl)
library(flora)
library(data.table)
library(rnaturalearth)
library(plotly)
library(hrbrthemes)
library(paletteer)
library(viridis)
library(scales)
library(patchwork)
library(showtext)
library(ggchicklet)
library(ggsflabel)
library(ggpubr)
font_add_google("Roboto Condensed") #just run once
